Split Thesaurus into provider/controller
